Fix code rewriting leads to incorrect output.
I reused a previous code rewriting technique instead of the current one.
Previously we were adding indirections before and after existing lines of code.
In this v1.0.2,  we replace messages and assigments with new messages, so everything becomes a message.
I don't know if the AST mapping will still hold when I will integrate that with the reflective method.