Modifiy the method modifier to create a completely new AST for  the instrumented method.
Originally we were using a shallow copy of the original AST.
So when modifying the copied AST it is possible that we are modifying nodes from the original AST.
This might be the cause of some problems if the original AST is recompiled at some point.
So now the visitor recreate all nodes of the AST when applying modifications (except for leaves).