Fix incorrect printing of a CompiledBlock source code when the latter is instrumented.
The code of a CompiledBlock is computed from the source code of its home method and its corresponding PC (computed with a lookup of the method literals).
Problem, the node we want to use to represent the CompiledBlock is in the not-instrumented method.
To return it, I added a bound between the original source code (ast node) and the instrumented CompiledBlock.